Restructuring and building delete crud  of several apis with full structuring api of BillingInfo

+++ Backend/ClothingManagment/src/main/java/edu/icet/clothify/service/impl/CartServiceImpl.java
@@ -45,6 +45,11 @@ public class CartServiceImpl implements CartService {
 
     @Override
     public void deleteCart(Long id) {
-        cartRepository.deleteById(id);
+//        cartRepository.deleteById(id);
+        if (cartRepository.existsById(id)){
+            cartRepository.deleteById(id);
+        }else {
+            throw new ResourceNotFoundException("card info not available for this id to delete: "+id);
+        }
     }
 }
